+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/state/channels/list/ItemState.kt
@@ -36,7 +36,7 @@ public sealed class ItemState {
     public data class ChannelItemState(
         val channel: Channel,
         val isMuted: Boolean = false,
-        val typingUsers: List<User>,
+        val typingUsers: List<User> = emptyList(),
     ) : ItemState() {
         override val key: String = channel.cid
     }