+++ stream-chat-android-compose/src/test/kotlin/io/getstream/chat/android/compose/ui/SnapshotTest.kt
@@ -25,30 +25,24 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalInspectionMode
 import app.cash.paparazzi.Paparazzi
-import io.getstream.chat.android.client.ChatClient
-import io.getstream.chat.android.client.audio.AudioPlayer
-import io.getstream.chat.android.compose.ui.attachments.StreamAttachmentFactories
 import io.getstream.chat.android.compose.ui.theme.ChatTheme
-import org.mockito.kotlin.doReturn
-import org.mockito.kotlin.mock
+import org.junit.Rule
 
-internal abstract class BaseComposeTest {
+internal interface SnapshotTest : ComposeTest {
 
-    abstract fun basePaparazzi(): Paparazzi
+    @get:Rule
+    val paparazzi: Paparazzi
 
     fun snapshot(
         isInDarkMode: Boolean = false,
         composable: @Composable () -> Unit,
     ) {
-        basePaparazzi().snapshot {
+        paparazzi.snapshot {
             CompositionLocalProvider(
                 LocalInspectionMode provides true,
             ) {
                 ChatTheme(
                     isInDarkMode = isInDarkMode,
-                    attachmentFactories = StreamAttachmentFactories.defaultFactories(
-                        getChatClient = this::getChatClient,
-                    ),
                 ) {
                     Box(modifier = Modifier.background(ChatTheme.colors.appBackground)) {
                         composable.invoke()
@@ -59,17 +53,13 @@ internal abstract class BaseComposeTest {
     }
 
     fun snapshotWithDarkMode(composable: @Composable () -> Unit) {
-        basePaparazzi().snapshot {
+        paparazzi.snapshot {
             CompositionLocalProvider(
                 LocalInspectionMode provides true,
             ) {
-                val attachmentFactories = StreamAttachmentFactories.defaultFactories(
-                    getChatClient = this::getChatClient,
-                )
                 Column {
                     ChatTheme(
                         isInDarkMode = true,
-                        attachmentFactories = attachmentFactories,
                     ) {
                         Box(modifier = Modifier.background(ChatTheme.colors.appBackground)) {
                             composable.invoke()
@@ -77,7 +67,6 @@ internal abstract class BaseComposeTest {
                     }
                     ChatTheme(
                         isInDarkMode = false,
-                        attachmentFactories = attachmentFactories,
                     ) {
                         Box(modifier = Modifier.background(ChatTheme.colors.appBackground)) {
                             composable.invoke()
@@ -89,7 +78,7 @@ internal abstract class BaseComposeTest {
     }
 
     fun snapshotWithDarkModeRow(composable: @Composable () -> Unit) {
-        basePaparazzi().snapshot {
+        paparazzi.snapshot {
             CompositionLocalProvider(
                 LocalInspectionMode provides true,
             ) {
@@ -105,14 +94,6 @@ internal abstract class BaseComposeTest {
                         }
                     }
                 }
-            }
-        }
-    }
-
-    private fun getChatClient(): ChatClient {
-        return mock {
-            on { getCurrentOrStoredUserId() } doReturn "mockedUser"
-            on { audioPlayer } doReturn mock<AudioPlayer> {
             }
         }
     }